Unbreak macos upload script

+++ integration/macos/publish.sh
@@ -3,16 +3,20 @@
 set -e
 
 mkdir -p "${HOME}/.ssh" && chmod 700 "${HOME}/.ssh"
-echo "${PACKAGE_UPLOAD_SSH_KEY}" > "${HOME}/.ssh/id_ed25519"
+if [ -f "${PACKAGE_UPLOAD_SSH_KEY_FILE}" ]; then
+    cp "${PACKAGE_UPLOAD_SSH_KEY_FILE}" "${HOME}/.ssh/id_ed25519"
+else
+    echo "${PACKAGE_UPLOAD_SSH_KEY}" > "${HOME}/.ssh/id_ed25519"
+fi
 chmod 400 "${HOME}/.ssh/id_ed25519"
+
+set -ex
 
 cat <<EOF >"${HOME}/.ssh/config"
 Host upload-packages.edgedb.com
     Port 2224
     StrictHostKeyChecking no
 EOF
-
-set -ex
 
 dest="artifacts"
 key=""
@@ -25,14 +29,14 @@ fi
 
 cd "${dest}"
 list=$(mktemp)
-rellist=$(python3 -c "import os.path; print(os.path.relpath(\"${list}\"))")
 batch=$(mktemp)
-find . -type f > "${list}"
+find . -type f -print | sed "s|^\./||" > "${list}"
+chmod g+rw "${list}"
 
 cat <<EOF >${batch}
-put -r * ./incoming/
-put ${rellist} ./incoming/triggers/upload${key}.list
+put -r * incoming/
+put ${list} incoming/triggers/upload${key}.list
 EOF
-
 sftp -b "${batch}" uploader@upload-packages.edgedb.com
+
 rm "${list}" "${batch}"
